% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{generate_func_var_roxygen_comments_from_file}
\alias{generate_func_var_roxygen_comments_from_file}
\title{Parse and convert text generated by doc_vars_in_this_func() from file}
\usage{
generate_func_var_roxygen_comments_from_file(infile, sinkFilePath,
  visual_separator = "\\n\\n")
}
\arguments{
\item{infile}{character string giving path and file name for input file}

\item{sinkFilePath}{character string giving path and file name for output file}

\item{visual_separator}{character string to write between the end of the
previous block and the new block to make it easier to pick out where
one ends and the other begins}
}
\value{
Returns nothing
}
\description{
This is a wrapper function for
\code{\link{generate_func_var_roxygen_comments_from_vec}} to load the
strings to be parsed from a file into the vector of strings.
}
\seealso{
\code{\link{generate_func_var_roxygen_comments_from_vec}}
}
