% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{generate_func_var_roxygen_comments_from_vec}
\alias{generate_func_var_roxygen_comments_from_vec}
\title{Parse and convert text generated by doc_vars_in_this_func() from vec of strings}
\usage{
generate_func_var_roxygen_comments_from_vec(all_data,
  visual_separator = "\\n\\n")
}
\arguments{
\item{all_data}{vector of character strings containing output from
doc_vars_in_this_func()}

\item{visual_separator}{character string to write between the end of the
previous block and the new block to make it easier to pick out where
one ends and the other begins}
}
\value{
Returns nothing
}
\description{
Parse and convert text generated by doc_vars_in_this_func() from vec of strings
}
\seealso{
\code{\link{generate_func_var_roxygen_comments_from_file}}
}
