% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{generate_func_var_roxygen_comments_from_file}
\alias{generate_func_var_roxygen_comments_from_file}
\title{Parse and convert text generated by doc_vars_in_this_func() from file}
\usage{
generate_func_var_roxygen_comments_from_file(infile, sinkFilePath)
}
\arguments{
\item{infile}{character string giving path and file name for input file}

\item{sinkFilePath}{character string giving path and file name for output file}
}
\value{
Returns nothing
}
\description{
This is a wrapper function for
\code{\link{generate_func_var_roxygen_comments_from_vec}} to load the
strings to be parsed from a file into the vector of strings.
}
\examples{
\dontrun{
workdir = "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.completedTzarEmulation"
infile = file.path (workdir, "consoleSinkOutput.temp.txt")
outfile = file.path (workdir, "localvar_roxygen_comments.txt")
generate_func_var_roxygen_comments_from_file (infile, outfile)
}
}
\seealso{
\code{\link{generate_func_var_roxygen_comments_from_vec}}
}
